+++ src/SDSUtils.h
@@ -1,0 +1,9 @@
+#ifndef SDSUtils_H
+#define SDSUtils_H
+
+#include <Rcpp.h>
+
+std::string SignificanceStar(double p);
+Rcpp::IntegerVector RcppUnique(Rcpp::IntegerVector x);
+
+#endif // SDSUtils_H